rename parse to header because we will also generate headers from this file
